+timeleaf web interface

+++ src/main/java/com/work/test/service/OrderService.java
@@ -2,12 +2,15 @@ package com.work.test.service;
 
 import com.work.test.dao.BookEntity;
 import com.work.test.dao.BookRepository;
+import com.work.test.dao.CustomerEntity;
 import com.work.test.dao.CustomerRepository;
 import com.work.test.dao.OrderEntity;
 import com.work.test.dao.OrderRepository;
 import com.work.test.dto.Order;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,41 +26,80 @@ public class OrderService {
     @Autowired
     private BookRepository bookRepository;
 
-    public List<Order> getById(Integer id) {
-
-        return new ArrayList<>();
-    }
-
     public void deleteOrder(Integer id) {
 
-        orderRepository.deleteById(id);
+        OrderEntity entity = orderRepository.findById(id).orElse(null);
+        if (entity.getFinished() == false) {
+            orderRepository.deleteById(id);
+        }
     }
 
     public void setFinished(Integer id) {
 
+        OrderEntity entity = orderRepository.findById(id).orElse(null);
+        if (entity == null) {
+            return;
+        }
+        entity.setFinished();
+        entity.setStoppedAt(new Date(System.currentTimeMillis()));
+        orderRepository.saveAndFlush(entity);
     }
 
     public void updateOrder(Order order) {
 
+        OrderEntity entity = orderRepository.findById(order.getId()).orElse(null);
+        if (entity != null) {
+            if (entity.getFinished() != true) {
+                orderRepository.saveAndFlush(dtoToDao(order));
+            }
+        }
     }
 
     public Integer createOrder(Order order) {
 
         OrderEntity entity = dtoToDao(order);
+        entity.setStartedAt(new Date(System.currentTimeMillis()));
         return orderRepository.saveAndFlush(entity).getId();
+    }
+
+    public List<Order> findById(Integer id) {
+
+        if (id != null) {
+            List<Order> orderList = new ArrayList<>();
+            OrderEntity entity = orderRepository.findById(id).orElse(null);
+            if (entity != null) {
+                orderList.add(daoToDto(entity));
+            }
+            return orderList;
+        }
+
+        else {
+            return getAll();
+        }
+    }
+
+    public List<Order> getAll() {
+        return orderRepository
+                .findAll()
+                .stream().map(entity -> { return daoToDto(entity); })
+                .collect(Collectors.toList());
     }
 
     private Order daoToDto(OrderEntity entity) {
 
         Order order = new Order();
-        order.setId(entity.getCustomer().getId());
+        order.setId(entity.getId());
+        order.setCustomerId(entity.getCustomer().getId());
         order.setName(entity.getName());
         order.setFinished(entity.getFinished());
         order.setStartedAt(entity.getStartedAt());
         order.setStoppedAt(entity.getStoppedAt());
         entity.getBooks()
                 .stream()
-                .forEach( bookEntity -> { order.addBook(bookEntity.getId()); });
+                .forEach(bookEntity -> {
+                    order.addBook(bookEntity.getId());
+                });
+
         return order;
     }
 
@@ -67,7 +109,9 @@ public class OrderService {
         entity.setId(order.getId());
         entity.setName(order.getName());
         entity.setFinished(order.isFinished());
-        entity.setCustomer(customerRepository.findById(order.getCustomerId()).orElse(null));
+        Integer customerId = order.getCustomerId();
+        CustomerEntity customerEntity = customerRepository.findById(customerId).orElse(null);
+        entity.setCustomer(customerEntity);
         order.getBooks()
                 .stream()
                 .forEach(bookId -> {
@@ -76,7 +120,6 @@ public class OrderService {
                         entity.addBook(bookEntity);
                     }
                 });
-
         return entity;
     }
 }
